misc updates
git-svn-id: https://nipy.svn.sourceforge.net/svnroot/nipy/nipype/trunk@183 ead46cd0-7350-4e37-8683-fc4c6f79bf00

+++ nipype/utils/filemanip.py
@@ -69,9 +69,11 @@ def copyfiles(originalfile, newfile,copy=False):
          
     Returns
     -------
-    output : dict
-        dictionary holding 'out', 'err', 'returncode'
+    None
+    
     """
+    if os.path.exists(newfile):
+        return
     if os.name is 'posix' and not copy:
         os.symlink(originalfile,newfile)
     else:
